Add files via upload
This repository contains the raw data, the analyzed and visualized data using Excel, the screenshot image of analyzed data and a "ReadMe" file containing all the processes involved in cleaning, analysis and generating insights on "Fashion Retail Items" in some part of the world over a period of time. It clearly shows the rate at which fashion items are purchased, revenue generated, and other relevant information.