Rename README.md to README.md.
This file clearly shows all the step-by-step processes and approach used in analyzing and visualizing the data using Excel on "Fashion Retail Items" in some part of the world over a period of time. It clearly shows the rate at which fashion items are purchased, revenue generated, and other relevant information.